Extract a common logger the analyzer framework
In development of the Rust FE logging is useful in debugging. Instead of
rolling our own logger it became clear the loggers part of analyzer and jit
projects could be extracted and made generic so they could be reused in
other projects.

To test this patch make check-jit was invoked, for analyzer the following
flags were used -fanalyzer -fdump-analyzer -fanalyzer-verbosity=4.

gcc/ChangeLog:

2021-03-8  Philip Herron <philip.herron@embecosm.com>

        * logging.h: added new generic logger based off analyzer's logger
        * logging.c: Likewise

gcc/analyzer/ChangeLog:

2021-03-8  Philip Herron <philip.herron@embecosm.com>

        * analyzer-logging.h: has been extract out to gcc/logging.h
        * analyzer-logging.c: Likewise